Downgrade pip to work around poise-python issue
There is a bug in the poise-python cookbook because it depends on
internal features of pip and these internals have changed in pip 18.1,
see [0]. We have a workaround in the openstack-common cookbook proposed,
but that doesn't work if the server has pip 18.1 pre-installed, as seems
to be the case for the infra-provided CI image. So as a further
workaround we need to downgrade the pre-installed pip version for the
time being.

[0] https://github.com/poise/poise-python/issues/133

Depends-On: https://review.openstack.org/610427
Change-Id: I52514691d3f271579628b62539222f7307957ba0