Chef 16 support
This adds support for testing both Chef/Cinc 15 and Chef/Cinc 16 which
are still supported upstream. For now this disables voting on 16 as we
have some known issues that need to be worked out first.

In addition, this updates to Chef Workstation 20.8.111 which includes
Chef 16 for spec testing. For now, let's only worry about doing unit
testing against Chef 16 and worry about integration testing to catch any
issues with Chef 15.

Also remove installation of the fog-openstack gem as it's now included
in Chef Workstation by default.

Change-Id: I9b99837c4603496a2cacde34a73c4e27109620c2
Signed-off-by: Lance Albertson <lance@osuosl.org>