Disable SELinux for integration testing on RHEL only
We currently don't fully support SELinux so we need to disable it for now until
we get it fixed properly. Make this a variable so that we can change dynamically
later if we need to for testing purposes.

Temporarily disable voting until we get things working properly again. Update
git URI for bind cookbook to use upstream master.

Change-Id: Ib6dec3f6ea1cbc13d9c82d372b4f65f689adc0bc
Signed-off-by: Lance Albertson <lance@osuosl.org>